Issue #149: RapidSVN will not compile on AMD64
Patch contributed by Tristan Keen:

* include/svncpp/datetime.hpp, src/svncpp/datetime.cpp 
  Added new files/class to aid type identity. 
 
* include/svncpp/revision.hpp, src/svncpp/revision.cpp 
  Changed constructor to use DateTime rather than apr_time_t directly. 
 
* src/rapidsvn_frame.cpp 
  (RapidSvnFrame::~RapidSvnFrame, RapidSvnFrame::InitFolderBrowser): 
    Standardised on longs for bookmark counts & their printf specifiers. 
 
* src/filelist_ctrl.hpp 
  Defined IndexArray as containing longs correctly. 
 
* src/diff_dlg.cpp 
  (RevisionPanel::GetRevision): Switched to using DateTime 
  (RevisionPanel::SetRevision): Corrected svn_revnum_t printf format 
 
* src/diff_action.cpp 
  (DiffAction::Data::getFile): Corrected svn_revnum_t printf format 
 
* src/svncpp/client_cat.cpp 
  (openTempFile): Corrected svn_revnum_t printf format 

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7019 65390229-12b7-0310-b90b-f21a5aa7ec8e

+++ src/rapidsvn_frame.cpp
@@ -81,7 +81,7 @@ const static char ConfigWidth[] = "/MainFrame/Width";
 const static char ConfigHeight[] = "/MainFrame/Height";
 const static char ConfigSplitterHoriz[] = "/MainFrame/SplitterHoriz";
 const static char ConfigSplitterVert[] = "/MainFrame/SplitterVert";
-const static char ConfigBookmarkFmt[] = "/Bookmarks/Bookmark%d";
+const static char ConfigBookmarkFmt[] = "/Bookmarks/Bookmark%ld";
 const static char ConfigBookmarkCount[] = "/Bookmarks/Count";
 
 const static char TraceMisc[] = "tracemisc";
@@ -467,9 +467,9 @@ RapidSvnFrame::~RapidSvnFrame ()
 
 
   // Save the bookmarks contents
-  size_t item;
-  const size_t itemCount = m_folder_browser->GetBookmarkCount ();
-  pConfig->Write (ConfigBookmarkCount, (long)itemCount);
+  long item;
+  const long itemCount = m_folder_browser->GetBookmarkCount ();
+  pConfig->Write (ConfigBookmarkCount, itemCount);
   for (item = 0; item < itemCount; item++)
   {
     wxString key;
@@ -838,7 +838,7 @@ RapidSvnFrame::InitFolderBrowser ()
   wxString key;
   wxString bookmark;
 
-  int item, count;
+  long item, count;
   pConfig->Read (ConfigBookmarkCount, &count, 0);
   for (item = 0; item < count; item++)
   {
